+++ extensions/capacities/CHANGELOG.md
@@ -4,5 +4,6 @@
 
 - in `Save Daily Note` you can choose whether to append timestamp or not
 - in `Save Daily Note` show Error view when Store fails to load
+- in `Search Content` show Error view when Store fails to load
 
 ## [Initial Version] - 2024-02-27